qa/workunits/fs/misc: remove data pool cleanup
This cleanup is at the very least incorrect as it can cause the MDS to throw
read-only errors because the data pool is removed before it can write out
backtraces.

We've not yet finalized a truly safe workflow to remove a data pool -- even
flushing the MDS journals first may not be enough (considering a large purge
queue).

Fixes: https://tracker.ceph.com/issues/70919
Signed-off-by: Patrick Donnelly <pdonnell@ibm.com>