crimson/osd/osd_operations/pg_advance_map: Add splitting as a function
As we initiate pg splitting as part of the PGAdvanceMap workflow, it is not required
to maintain it as a separate osd_operation.
A new function in PGAdvanceMap - split_pg(), will now take care of the splitting workflow
if we detect split children in an OSD map.

Since we do not follow the same queuing system as classical OSD in crimson, we will not
need to maintain pg_num_history. This makes the splitting check simpler.
With most of the splitting code being part of PGAdvanceMap, it makes sense to have the
splitting check there as well and leave broadcast_map_to_pgs untouched.

Signed-off-by: Aishwarya Mathuria <amathuri@redhat.com>