rbd-mirror: fix missing mirror group status
The recent changes to the GroupReplayer Bootstrap caused the
mirror group status for a newly enabled group to not be written to
the rbd_mirroring object on the secondary for upto 10 seconds.
The cls_rbd group_status_set did not write the status to disk if
the group does not exist or if its state is not enabled. This has been
modified to write the status even if the status is creating.

Signed-off-by: N Balachandran <nithya.balachandran@ibm.com>