osd: overaggressive assert in read_log_and_missing with optimized EC pool
read_log_and_missing is called during OSD initializaiton to sanity check
the PG log. One of its checks is too agressive for an optimized EC pool
where because of a partial write there can be a log entry but no update
to the object on this shard (other shards will have been updated). The
fix is to skip the checks when the log entry indicates this shard was
not updated.

Only affects pool with allow_ec_optimizations flag on.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>
(cherry picked from commit c8739a2bdb0bf523402d85517a7fce39d445eef5)