doc/radosgw: Promptify CLI commands and fix formatting in layout.rst
Move two example CLI command blocks to use a bash prompt
block with the correct privileged prompt instead of
hardcoded unprivileged prompt.

Don't use the directory "bin/" in command "bin/radosgw-admin"
especially when the next commands also don't use it.

Fix 2 instances of double-backtick missing the second
backtick that caused the rendered doc to be jumbled.

Do not use _emphasis_ that renders as underscores, instead
use RST syntax for italic for the word.

Signed-off-by: Ville Ojamo <14869000+bluikko@users.noreply.github.com>