os/bluestore: Improve bluestore_debug_omit_block_device_write
Reused pool opt CSUM_MIN_BLOCK that is not used.
When it is set to 1, it works as bluestore_debug_omit_block_device_write,
but only for the one pool.

Also, reading from such pool always returns 0.

The rationale is to allow produce fragmented metadata faster.

One cannot use bluestore_debug_omit_block_device_write for it,
because OSD complains on missing osdmaps and superblock.

Signed-off-by: Adam Kupczyk <akupczyk@ibm.com>