doc/mgr: edit administrator.rst
Edit doc/mgr/administrator.rst. Add prompts and correct some of the
inelegant English present in the file.

This commit is part of a project to separate out the twenty-five files
that were committed to https://github.com/ceph/ceph/pull/62782.

Credit for this initiative should go to Ville Ojamo.

Signed-off-by: Zac Dover <zac.dover@proton.me>