librbd/mirror/PromoteRequest: return EINVAL
... instead of ENOENT when mirroring is disabled for the image.

The PromoteRequest async state machine returns ENOENT when mirroring is
disabled for the image. Instead, make it return EINVAL similar to
DemoteRequest's behavior, which is more appropriate. This also causes
the public facing C, C++, and Python APIs that promote an image
to return EINVAL or raise an equivalent exception when mirroring is
disabled for the image.

Signed-off-by: Ramana Raja <rraja@redhat.com>