crimson/osd/osd_operations/pg_advance_map: Add check for splitting
As we do not follow the same queuing system as classical OSD in crimson, we will not
need to maintain pg_num_history. This makes the splitting check simpler.
With most of the splitting code being part of PGAdvanceMap, it makes sense to have the
splitting check there as well and leave broadcast_map_to_pgs untouched.

Signed-off-by: Aishwarya Mathuria <amathuri@redhat.com>