doc/mgr: edit cli_api.rst
Edit doc/mgr/cli_api.rst as part of the project to determine where the
error is in https://github.com/ceph/ceph/pull/62782 that prevents the
Jenkins tests from passing.

This is a change to one of twenty-five files in
https://github.com/ceph/ceph/pull/62782 and this commit represents one
of what will be at least twenty-five other commits made to track this
error down.

Signed-off-by: Zac Dover <zac.dover@proton.me>