doc/mgr: edit dashboard.rst
Edit doc/mgr/crash.rst. Add prompts.

This changes eighty-nine prompts. Because this makes so many changes,
all other edits included in https://github.com/ceph/ceph/pull/63255 will
be made in a separate commit. This done for the sake of the patience of
the reviewers (probably Anthony, if history is any guide).

This commit is part of a project to separate out the twenty-five files
that were committed to https://github.com/ceph/ceph/pull/63255.

Signed-off-by: Zac Dover <zac.dover@proton.me>