pybind/mgr: Fix missing empty line in documentation in user.py
The inline method documentation needs an empty line between description
and metadata.

Otherwise automethod from docs renders the metadata as text together
with the description instead of rendering it as function metadata, e.g.
parameters and return value.

This PR adds missing empty line in inline documentation in user.py and
that seems to fix the documentation.

Signed-off-by: Ville Ojamo <14869000+bluikko@users.noreply.github.com>