osd: EC Optimizations fix missing call to partial_write
When a shard is backfilling and it receives a log entry where the
transaction is not applied it can skip the roll forward by
immediately advancing crt. However it is still necessary to
call partial_write in this scenario to keep the pwlc information
up to date.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>
(cherry picked from commit ddc306255868f26ae0a3951710ef18207fff9b30)