qa/standalone/scrub: extra delay in osd-scrub-test.sh
Modify TEST_just_deep_scrubs():
The test turns 'no-deep' off, and expects the scrub to be
concluded and reported within a short period. But the PG
already had a chance to try scrubbing (and fail due to
the 'no-deep'), and its 'not-before'
was pushed out by osd_scrub_retry_after_noscrub.

The fix is to reduce the value of osd_scrub_retry_after_noscrub
for this test, and to increase the timeout for the test itself.

Fixes: https://tracker.ceph.com/issues/71404

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>
(cherry picked from commit cb91678ab09723088b10525952333b20810e8eb9)