mon/MgrStatMonitor.cc: do not update score when disabled
This commit adds changes to ensure the availability score
tracking is not updated when the feature is disabled. We
will preserve the score calculated before the feature is
turned off and start updating it again when the feature
is enabled.

Fixes: https://tracker.ceph.com/issues/71494

Signed-off-by: Shraddha Agrawal <shraddhaag@ibm.com>
(cherry picked from commit 017c9b9f4fa0d3286d52b7acb7df912327b1f836)