objecter: request OSDMap after idle ticks
If the objecter goes two ticks without receiving an OSDMap, it may
fall behind on incremental updates.  Call `maybe_request_osdmap()` in
the tick handler whenever we detect that two consecutive ticks have
elapsed without a fresh map, so we minimize the time needed to catch
up on missed changes.

Fixes: https://tracker.ceph.com/issues/71261
Signed-off-by: Nitzan Mordechai <nmordec@redhat.com>