mon/MgrStatMonitor: reset pool_availability when config option is toggled
This commit resets the last_uptime and last_downtime for each pool
when the track availability score feature is toggled from disbaled
to enabled. This is done to ensure when the feature is enabled
again and we start to track availability, we do not consider the
time elapsed in between (the duration for which the feature was
disabled) in the score calculation.

All other toggle states for the config option will not have any
effect on the score.

Signed-off-by: Shraddha Agrawal <shraddha.agrawal000@gmail.com>