rgw: make sure max_objs_per_shard is appropriate in debugging scenarios
When we have a versioned bucket, we reduce max_objs_per_shard by a
factor of 3 to account for the extra bucket index entries required in
such buckets. And during debugging, we may want to induce early
resharding by setting max_objs_per_shard to an artificially low
value. Combined, that math could result in max_objs_per_shard with a
value of 0 that would cause a division by zero crash. This fixes that.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>