snaptrim: optimize snapshot deletion performance and CPU utilization
Problem:
When deleting snapshots with over 10k objects, due to the linear scanning of object entries, the deletion time may be as long as 400 seconds.
In addition, the accumulation of idle empty snapshots will cause the entire storage node to maintain a CPU usage rate of over 60% continuously.

Solution:
A large amount of delay in the current snapshot deletion is spent on RocksDB seek.
Optimize the query logic by saving the last scanned object as a prefix for the next query.
This avoids full scans in each iteration.

Introduce dynamic sleep mechanism controlled by new config
  osd_next_snap_trim_sleep (default: 0)

Verification:
| Scenario            | Before    | After     |
|---------------------|-----------|-----------|
| 10k objects         | 400s      | 8s        |
| 1k empty snaps      | 60% CPU   | <5% CPU   |

Signed-off-by: wzc <wangzechen@inspur.com>