doc/mgr: edit modules.rst.
Edit doc/mgr/modules.rst. Improve some of the English.

This is part of a project to separate out the twenty-five files that
were committed to https://github.com/ceph/ceph/pull/62782.

This commit and the PR of which it is a part includes the changes made
in https://github.com/ceph/ceph/pull/62782 and some corrections to
glaring grammatical errors. However, this file could use a careful
grammar pass. As it currently stands, it contains run-on sentences and
English that is acceptable in conversation but not acceptable in written
communication.

Signed-off-by: Zac Dover <zac.dover@proton.me>