include/ceph_assert: drop ceph_assert_always()
Many years ago there was an idea to find the non-critical occurrences
of `ceph_assert()`, replace them with the then-new `ceph_assert_always()`
and finally let the remaining to be compiled-out in release builds.

Since then the adoption of `ceph_assert_always()` (maybe apart of crimson)
turned out be negligible while preserving `ceph_assert()` always compiled,
IMHO, has saved our necks so many times that the penalty on performance
would have need to be **tremendus** to justify the transition.

However, the lengthy limbo state is confusing. We should either finish
the transition (but how to quickly judge which occurences are _ciritical_?)
or scrap it. This commit does the latter.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>