rbd-mirror: avoid altering local group snap vector elements
Introduce m_create_group_snaps which hold snaps which are in creation phase
and take them to complete. Now the m_local_group_snaps is dedicated to
ListSnapshotsRequest() locally and is unaltered.

Note, this patch fixes the crash, that can be seen with user snap
created, then remove and then created with the same name.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>