mon: provide emergency mechanism to use mon keyring
If they key is lost for the `mon.` credential, it's very inconvenient to get it
out of the "auth" database in the mon store. So, allow the operator to create a
new keyring for the mons and use it instead to get mons in quorum again.

Signed-off-by: Patrick Donnelly <pdonnell@ibm.com>
Change-Id: 46a7a5e3-a7e7-4a25-ad4e-710d5399c2a8
(cherry picked from commit 8d8eb5445b9ef393767225f0eead1b6a7ef59bf4)