auth: check service key is valid before decryption
CryptoKey::empty is the correct mechanism to check for an invalid key (and this
is codified elsewhere, fixed in this commit). Decryption would fail with an
abort if the key handler was unset. This would happen after rotating the "mon."
key and then restarting one of the mons.

Signed-off-by: Patrick Donnelly <pdonnell@ibm.com>
Change-Id: cc810b37-c7d2-4417-aee4-f9444e966448
(cherry picked from commit fe5e6d50b2d8bdc923f8bc4cfff2cabe955f6196)