mgr/dashboard: add a cobertura xml file workaround variable
Add an environment variable REWRITE_COVERAGE_ROOTDIR that
changes the "hardcoded" path in the cobertura-coverage.xml file.
This can be used to map the paths used in a container build to
the paths known to a jenkins job (or whatever else you want to
do with the file).

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit f80dad2e0adfb5bca33fab292728f5c085ca3a78)