mgr/cephadm: add DaemonDeployQueue class
Basically a light wrapper around a daemon type to list
of DaemonPlacement (class scheduler produces for adding
a daemon). The plan is to set up a global instance of
this for the whole cephadm mgr module like most of the other
classes in inventory.py and we can queue up daemons we want
to deploy when applying specs for actual deployment later.

Signed-off-by: Adam King <adking@redhat.com>