rgw/logging: return the last object name that was actually comitted
when comitting a pending object that was never created we should
not reply the object name as the name of the comitted object.
instead, we should return the name of the object that was actuaslly
comitted.

Fixes: https://tracker.ceph.com/issues/71219

Signed-off-by: Yuval Lifshitz <ylifshit@ibm.com>
(cherry picked from commit efe4473000e1a4510549c9e39ce984e8f3bb1db6)