DNM: mgr/vol: don't unlink meta symlink until new incarnation is created
So far the meta symlink is deleted as soon as the subvolume is removed,
regardless of whether snapshots are retained. This leads to a major
issue with v3.

Unlinking the meta symlink is especially a problem in case of subvolume
deletion with snapshots retained. It breaks subvolume discovery code
(absence of meta file implies legacy subvolume acc to current code) and
the code for opening v2 and v3 subvolumes (subvolume type is needed to
check if the op for which subvolume is to be opened is allowed and
absence of meta file makes it impossible to subvolume type).

With this patch the behaviour is to not delete the meta file until
subvolume is being deleted along with all of the snapshots, if they
exist. This allows to operate on this subvolume as meta file is still
present.

If new incarnation is created for a subvolume with retained snapshots,
the meta symlink needs to be modofied to new incarnation's meta file
instead of old incarnation's meta file.

Signed-off-by: Rishabh Dave <ridave@redhat.com>