crimson-osd: centralize parameter processing
Some of the OSD seastar application options (mkkey, prometheus, etc) are not
presented in the `crimson-osd --help`.
By moving the supported options into a dedicated `crimson_options_t` struct and
collecting them earlier in `get_early_config` we could present them in --help calls.

Any options which are not part of this struct would be pushed into
`unrecognized_options` to later be parsed by `_get_early_config` **same**
as before this change.

`crimson_options` don't need to be passed through the pipe and could be directly
added to the returned result from `get_early_config` from the parent process.

Fixes: https://tracker.ceph.com/issues/71573
Signed-off-by: junxiang Mu <1948535941@qq.com>