rgw/sal: introduce DataProcessorFactory
This is a factory for DataProcessor to enable reading the data by
RGWGetObj_Filter and pass it to be processed by the writer that is
defined by set_writer().
With that, we can have filters applied to the data before writing.
That can be either filter on read or filter on write.

Signed-off-by: Seena Fallah <seenafallah@gmail.com>