rgw/sal: introduce DataProcessorFactory
This is a factory for DataProcessor to enable reading the data by
RGWGetObj_Filter and pass it to be processed by the writer that is
defined by set_writer().
With that, we can have filters applied to the data before writing.
That can be either filter on read or filter on write.

Co-authored-by: Marcus Watts <mwatts@redhat.com>
Signed-off-by: Seena Fallah <seenafallah@gmail.com>
(cherry picked from commit 95d3bc70d1dc31f7092607b3083e443e1084613c)