crimson: Use osd_objectstore instead of using crimson_osd_objectstore
Originally the idea with introducing a crimson dedicated
configurable was to allow changing the default object store to
SeaStore without affecting the Classic option. Moreover, if we
were to consider hybrid clusters, different OSD types would require
two different dedicated options.

Given that cephadm (and possibly other orchestrators) use
`--osd-obejctstore` to set the backend - we would prefer to
retain (at least for now) this behavior to not break the
**existing** cluster deployment methods.

Note: A follow up PR would be pushed against cephadm to
      fully support Seastore as backend option.

Fixes: https://tracker.ceph.com/issues/71593
Signed-off-by: Mohit Agrawal <moagrawa@redhat.com>