python-common/deployment: add remote control sidecar config params
Add parameters to the smb service spec for configuring the remote
control sidecar. The feature always controls the deployment (or not) of
the sidecar and the ssl/tls config params will be used to set up (m)TLS
for the sidecar.

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit 09da84fe1879ddb1752fceee4746880477cf0c04)

Conflicts:
	src/pybind/mgr/cephadm/tests/test_services.py